fixed 同步汽车版本最新源码

+++ backend/fuint-repository/src/main/java/com/fuint/repository/mapper/TAccountDutyMapper.java
@@ -13,8 +13,8 @@ import java.util.List;
  */
 public interface TAccountDutyMapper extends BaseMapper<TAccountDuty> {
 
-   List<Long> getDutyIdsByAccountId(long accountId);
+   List<Integer> getDutyIdsByAccountId(Integer accountId);
 
-   void deleteDutiesByAccountId(long accountId);
+   void deleteDutiesByAccountId(Integer accountId);
 
 }
